# Improve performance in sql.activerecord subscription callback