@technicalpickles 2023-08-07
Snuck another small improvement: Use SHA256 with is marginally faster than SHA1

Benchmark:

```ruby
require 'benchmark/ips'
require 'digest'

def refactored
  query_caller = caller
  location_key = Digest::SHA1.hexdigest(query_caller.join)
  query_caller
end

def refactored_sha2
  query_caller = caller
  location_key = Digest::SHA256.hexdigest(query_caller.join)
  query_caller
end

Benchmark.ips do |x|
  x.config warmup: 0, times: 100
  x.report("refactored") { refactored }
  x.report("refactored_sha2") { refactored_sha2 }
  x.compare!
end
```

Results:

```
Calculating -------------------------------------
          refactored    272.508k (±16.2%) i/s -      1.157M in   4.794913s
     refactored_sha2    279.210k (±16.6%) i/s -      1.184M in   4.793971s

Comparison:
     refactored_sha2:   279210.1 i/s
          refactored:   272507.7 i/s - same-ish: difference falls within error
```