@charkost 2023-08-09
Nice perf improvements @technicalpickles, thanks!